add bot config fix up some condition because they're really messy fix whitelist bugs

+++ src/commands/unwhitelist.ts
@@ -2,8 +2,10 @@ import {
   CommandInteraction,
   SlashCommandBuilder,
   EmbedBuilder,
+  GuildMemberRoleManager,
 } from "discord.js";
 import { PrismaClient } from "@prisma/client";
+import config from "../../botConfig.json";
 const prisma = new PrismaClient();
 
 export const data = new SlashCommandBuilder()
@@ -19,25 +21,37 @@ export const data = new SlashCommandBuilder()
 export async function execute(interaction: CommandInteraction) {
   try {
     const user = interaction.options.getUser("user");
-    if (user) {
-      const databaseUser = await prisma.user.findUnique({
-        where: {
-          id: parseInt(user.id),
-        },
-      });
-      if (databaseUser) {
-        await prisma.user.delete({
+    //* Check if both users exists.
+    if (user && interaction.member) {
+      let runnerRole = (interaction.member.roles as GuildMemberRoleManager)
+        .cache;
+      //* Check if the member running the command has the correct role.
+      if (runnerRole.some((role) => config.whitelistRoles.includes(role.id))) {
+        const databaseUser = await prisma.user.findUnique({
           where: {
             id: parseInt(user.id),
           },
         });
-        const successEmbed = new EmbedBuilder()
-          .setTitle("User unwhitelisted.")
-          .setColor("Green");
-        return interaction.reply({ embeds: [successEmbed] });
+        //* Check if the user exists in the database.
+        if (databaseUser) {
+          await prisma.user.delete({
+            where: {
+              id: parseInt(user.id),
+            },
+          });
+          const successEmbed = new EmbedBuilder()
+            .setTitle("User unwhitelisted.")
+            .setColor("Green");
+          return interaction.reply({ embeds: [successEmbed] });
+        } else {
+          const errorEmbed = new EmbedBuilder()
+            .setTitle("User doesn't exist.")
+            .setColor("Red");
+          return interaction.reply({ embeds: [errorEmbed] });
+        }
       } else {
         const errorEmbed = new EmbedBuilder()
-          .setTitle("User doesn't exist.")
+          .setTitle("You don't have permission to run this command.")
           .setColor("Red");
         return interaction.reply({ embeds: [errorEmbed] });
       }
